requirements.txt
streamlit==1.28.2
langchain==0.1.14
langchain-openai==0.0.8
langchain-community==0.0.30
python-dotenv
langchain-groq